requirements.txt
azure-functions
azure-cosmos==4.6.0
